Revert "Adding New To-Do Items..."
